refactor: Lotto에 auto 프로퍼티 삭제
