+++ src/main/kotlin/lotto/controller/LottoController.kt
@@ -5,6 +5,7 @@ import lotto.domain.LottoPurchase.Companion.DEFAULT_PRICE
 import lotto.domain.Lottos
 import lotto.view.InputView
 import lotto.view.ResultView
+import lotto.vo.LottoPurchaseRequest
 
 class LottoController {
     private val inputView = InputView()
@@ -22,10 +23,11 @@ class LottoController {
     private fun buyLottos(budget: Int): Lottos {
         val manualAmount = inputView.inputManualPurchaseAmount()
         val manualLottosNumbers = inputView.inputManualLottos(manualAmount)
-        val lottoPurchase = LottoPurchase(budget, DEFAULT_PRICE, manualLottosNumbers)
+        val lottoPurchaseRequest = LottoPurchaseRequest(budget, DEFAULT_PRICE, manualLottosNumbers)
+        val lottoPurchase = LottoPurchase(lottoPurchaseRequest)
 
         val allLottos = lottoPurchase.purchaseManualAndAuto()
-        resultView.printPurchaseAmount(allLottos.numberOfManual(), allLottos.numberOfAuto())
+        resultView.printPurchaseAmount(lottoPurchase.manualAmount, lottoPurchase.autoAmount)
         resultView.printLottos(allLottos)
 
         return allLottos
