+++ src/main/kotlin/lotto/domain/LottoPurchase.kt
@@ -1,22 +1,26 @@
 package lotto.domain
 
 import lotto.domain.LottoErrorMessage.OVER_BUDGET
+import lotto.vo.LottoPurchaseRequest
 
 class LottoPurchase(
-    private val allBudget: Int,
-    private val priceOfLotto: Int,
-    private val manualLottosNumbers: List<List<Int>>,
+    private val lottoPurchaseRequest: LottoPurchaseRequest
 ) {
-    private val manualAmount = manualLottosNumbers.size
+    val manualAmount = lottoPurchaseRequest.manualLottosNumbers.size
+    var autoAmount = 0
+        private set
 
     init {
+        val priceOfLotto = lottoPurchaseRequest.priceOfLotto
+        val allBudget = lottoPurchaseRequest.allBudget
+
         require(manualAmount * priceOfLotto <= allBudget) { OVER_BUDGET }
     }
 
     fun purchaseManualAndAuto(): Lottos {
         val manualLottoList = manual()
 
-        val autoBudget = allBudget - (manualAmount * priceOfLotto)
+        val autoBudget = lottoPurchaseRequest.allBudget - (manualAmount * lottoPurchaseRequest.priceOfLotto)
         val autoLottoList = auto(autoBudget)
         return Lottos(manualLottoList.plus(autoLottoList))
     }
@@ -24,20 +28,22 @@ class LottoPurchase(
     private fun manual(): List<Lotto> {
         val lottoList = mutableListOf<Lotto>()
 
-        manualLottosNumbers.forEach {
+        lottoPurchaseRequest.manualLottosNumbers.forEach {
             lottoList.add(Lotto.manualCreate(it))
         }
         return lottoList
     }
 
     private fun auto(autoBudget: Int): List<Lotto> {
-        val amount = autoBudget / priceOfLotto
+        val amount = autoBudget / lottoPurchaseRequest.priceOfLotto
 
         val lottos = buildList {
             repeat(amount) {
                 add(Lotto.autoCreate())
             }
         }
+        autoAmount = amount
+
         return lottos
     }
