+++ src/main/kotlin/lotto/vo/LottoPurchaseRequest.kt
@@ -1,0 +1,7 @@
+package lotto.vo
+
+data class LottoPurchaseRequest(
+    val allBudget: Int,
+    val priceOfLotto: Int,
+    val manualLottosNumbers: List<List<Int>>
+)
